Build Order API

+++ src/main/java/com/dangquocdat/FoodOrdering/dto/cart/CartDto.java
@@ -17,7 +17,7 @@ public class CartDto {
 
     Long customerId;
 
-    Long totalPrice;
+    double totalPrice;
 
     List<CartItemDto> cartItems;
 }
